A start algorithm changed && failure state added

+++ SmartCity Core/src/main/java/be/uantwerpen/sc/repositories/MapPointRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 
 @Repository
 public interface MapPointRepository extends CrudRepository<MapPoint, Long> {
-    MapPoint findById(long id);
+    MapPoint findById(int id);
     MapPoint findByIdAndMap(int id, Map map);
     MapPoint findByPointIdAndMap_Id(int pointId, int mapId);
     List<MapPoint> findAllByMap_Id(int mapId);
